+++ januscloud/proxy/dao/mem_videocall_user_dao.py
@@ -15,20 +15,22 @@ class MemVideoCallUserDao(object):
         else:
             return None
 
-    def del_by_username(self, username):
-        self._users_by_name.pop(username, None)
+    def remove(self, videocall_user):
+        mem_user = self._users_by_name.get(videocall_user.username)
+        if mem_user and mem_user.handle is not videocall_user.handle:
+            # videocall_user has been replaced, just return
+            return
+        self._users_by_name.pop(videocall_user.username, None)
 
-    def add(self, video_call_user):
-        if video_call_user.username in self._users_by_name:
-            raise JanusCloudError('videocall user {} already in repo'.format(video_call_user.name), JANUS_ERROR_CONFLICT)
-        self._users_by_name[video_call_user.username] = copy.copy(video_call_user)
-
+    def add(self, videocall_user):
+        self._users_by_name[videocall_user.username] = copy.copy(videocall_user)
 
     def update(self, videocall_user):
         org_videocall_user = self._users_by_name.get(videocall_user.username)
         if not org_videocall_user:
-            raise JanusCloudError('server {} NOT found'.format(videocall_user.username), JANUS_ERROR_NOT_FOUND)
-        org_videocall_user.__dict__.update(videocall_user.__dict__)
+            self._users_by_name[videocall_user.username] = copy.copy(videocall_user)
+        else:
+            org_videocall_user.__dict__.update(videocall_user.__dict__)
 
     def get_username_list(self):
         return [video_call_user.username for video_call_user in self._users_by_name.values()]